Refactor SDFGState into its own file and finish other refactoring
