+++ dace/sdfg/__init__.py
@@ -1,27 +1,17 @@
-from dace.sdfg.sdfg import (
-    # sdfg.py
-    SDFG,
-    InterstateEdge,
-    SDFGState,
+from dace.sdfg.sdfg import SDFG, InterstateEdge
 
-    # graph.py
-    SubgraphView,
+from dace.sdfg.state import SDFGState
 
-    # scope.py
-    ScopeSubgraphView,
+from dace.sdfg.scope import (scope_contains_scope, is_devicelevel,
+                             devicelevel_block_size, ScopeSubgraphView)
 
-    # replace.py
-    replace,
-    replace_properties)
+from dace.sdfg.replace import replace, replace_properties
 
 from dace.sdfg.utils import (has_dynamic_map_inputs, dynamic_map_inputs,
                              is_parallel, concurrent_subgraphs,
                              find_input_arraynode, find_output_arraynode,
                              trace_nested_access, is_array_stream_view,
                              local_transients)
-
-from dace.sdfg.scope import (scope_contains_scope, is_devicelevel,
-                             devicelevel_block_size)
 
 from dace.sdfg.validation import (InvalidSDFGError, InvalidSDFGNodeError,
                                   InvalidSDFGEdgeError,
