+++ dace/sdfg/state.py
@@ -1,0 +1,1411 @@
+""" Contains classes of a single SDFG state and dataflow subgraphs. """
+
+import collections
+import copy
+from dace import dtypes, memlet as mm, serialize, subsets as sbs, symbolic
+from dace.sdfg import nodes as nd
+from dace.sdfg.graph import OrderedMultiDiConnectorGraph, MultiConnectorEdge
+from dace.sdfg.propagation import propagate_memlet
+from dace.sdfg.validation import validate_state
+from dace.properties import (Property, DictProperty, SubsetProperty,
+                             SymbolicProperty, CodeBlock, make_properties)
+from inspect import getframeinfo, stack
+from typing import Any, Dict, Optional, List, Set, Tuple, Union
+import warnings
+
+
+def _getdebuginfo(old_dinfo=None) -> dtypes.DebugInfo:
+    """ Returns a DebugInfo object for the position that called this function.
+        :param old_dinfo: Another DebugInfo object that will override the
+                          return value of this function
+        :return: DebugInfo containing line number and calling file.
+    """
+    if old_dinfo is not None:
+        return old_dinfo
+
+    caller = getframeinfo(stack()[2][0])
+    return dtypes.DebugInfo(caller.lineno, 0, caller.lineno, 0,
+                            caller.filename)
+
+
+class MemletTrackingView(object):
+    """ A mixin class that enables tracking memlets in directed acyclic multigraphs. """
+    def memlet_path(self,
+                    edge: MultiConnectorEdge) -> List[MultiConnectorEdge]:
+        """ Given one edge, returns a list of edges representing a path
+            between its source and sink nodes. Used for memlet tracking.
+
+            @note: Behavior is undefined when there is more than one path
+                   involving this edge.
+            :param edge: An edge within this state.
+            :return: A list of edges from a source node to a destination node.
+            """
+        result = [edge]
+
+        # Obtain the full state (to work with paths that trace beyond a scope)
+        state = self._graph
+
+        # If empty memlet, return itself as the path
+        if edge.src_conn is None and edge.dst_conn is None and edge.data.data is None:
+            return result
+
+        # Prepend incoming edges until reaching the source node
+        curedge = edge
+        while not isinstance(curedge.src, (nd.CodeNode, nd.AccessNode)):
+            # Trace through scopes using OUT_# -> IN_#
+            if isinstance(curedge.src, (nd.EntryNode, nd.ExitNode)):
+                if curedge.src_conn is None:
+                    raise ValueError(
+                        "Source connector cannot be None for {}".format(
+                            curedge.src))
+                assert curedge.src_conn.startswith("OUT_")
+                next_edge = next(e for e in state.in_edges(curedge.src)
+                                 if e.dst_conn == "IN_" + curedge.src_conn[4:])
+                result.insert(0, next_edge)
+                curedge = next_edge
+
+        # Prepend outgoing edges until reaching the sink node
+        curedge = edge
+        while not isinstance(curedge.dst, (nd.CodeNode, nd.AccessNode)):
+            # Trace through scope entry using IN_# -> OUT_#
+            if isinstance(curedge.dst, (nd.EntryNode, nd.ExitNode)):
+                if curedge.dst_conn is None:
+                    raise ValueError(
+                        "Destination connector cannot be None for {}".format(
+                            curedge.dst))
+                if not curedge.dst_conn.startswith("IN_"):  # Map variable
+                    break
+                next_edge = next(e for e in state.out_edges(curedge.dst)
+                                 if e.src_conn == "OUT_" +
+                                 curedge.dst_conn[3:])
+                result.append(next_edge)
+                curedge = next_edge
+
+        return result
+
+    def memlet_tree(self, edge: MultiConnectorEdge) -> mm.MemletTree:
+        """ Given one edge, returns a tree of edges between its node source(s)
+            and sink(s). Used for memlet tracking.
+
+            :param edge: An edge within this state.
+            :return: A tree of edges whose root is the source/sink node
+                     (depending on direction) and associated children edges.
+            """
+        propagate_forward = False
+        propagate_backward = False
+        if ((isinstance(edge.src, nd.EntryNode) and edge.src_conn is not None)
+                or
+            (isinstance(edge.dst, nd.EntryNode) and edge.dst_conn is not None
+             and edge.dst_conn.startswith('IN_'))):
+            propagate_forward = True
+        if ((isinstance(edge.src, nd.ExitNode) and edge.src_conn is not None)
+                or
+            (isinstance(edge.dst, nd.ExitNode) and edge.dst_conn is not None)):
+            propagate_backward = True
+
+        # If either both are False (no scopes involved) or both are True
+        # (invalid SDFG), we return only the current edge as a degenerate tree
+        if propagate_forward == propagate_backward:
+            return mm.MemletTree(edge)
+
+        # Obtain the full state (to work with paths that trace beyond a scope)
+        state = self._graph
+
+        # Find tree root
+        curedge = edge
+        if propagate_forward:
+            while (isinstance(curedge.src, nd.EntryNode)
+                   and curedge.src_conn is not None):
+                assert curedge.src_conn.startswith('OUT_')
+                cname = curedge.src_conn[4:]
+                curedge = next(e for e in state.in_edges(curedge.src)
+                               if e.dst_conn == 'IN_%s' % cname)
+        elif propagate_backward:
+            while (isinstance(curedge.dst, nd.ExitNode)
+                   and curedge.dst_conn is not None):
+                assert curedge.dst_conn.startswith('IN_')
+                cname = curedge.dst_conn[3:]
+                curedge = next(e for e in state.out_edges(curedge.dst)
+                               if e.src_conn == 'OUT_%s' % cname)
+        tree_root = mm.MemletTree(curedge)
+
+        # Collect children (recursively)
+        def add_children(treenode):
+            if propagate_forward:
+                if not (isinstance(treenode.edge.dst, nd.EntryNode)
+                        and treenode.edge.dst_conn
+                        and treenode.edge.dst_conn.startswith('IN_')):
+                    return
+                conn = treenode.edge.dst_conn[3:]
+                treenode.children = [
+                    mm.MemletTree(e, parent=treenode)
+                    for e in state.out_edges(treenode.edge.dst)
+                    if e.src_conn == 'OUT_%s' % conn
+                ]
+            elif propagate_backward:
+                if (not isinstance(treenode.edge.src, nd.ExitNode)
+                        or treenode.edge.src_conn is None):
+                    return
+                conn = treenode.edge.src_conn[4:]
+                treenode.children = [
+                    mm.MemletTree(e, parent=treenode)
+                    for e in state.in_edges(treenode.edge.src)
+                    if e.dst_conn == 'IN_%s' % conn
+                ]
+
+            for child in treenode.children:
+                add_children(child)
+
+        # Start from root node (obtained from above parent traversal)
+        add_children(tree_root)
+
+        # Find edge in tree
+        def traverse(node):
+            if node.edge == edge:
+                return node
+            for child in node.children:
+                res = traverse(child)
+                if res is not None:
+                    return res
+            return None
+
+        # Return node that corresponds to current edge
+        return traverse(tree_root)
+
+
+# TODO: Use mixin for SDFGState and ScopeSubgraphView for scope dict
+@make_properties
+class SDFGState(OrderedMultiDiConnectorGraph, MemletTrackingView):
+    """ An acyclic dataflow multigraph in an SDFG, corresponding to a
+        single state in the SDFG state machine. """
+
+    is_collapsed = Property(dtype=bool,
+                            desc="Show this node/scope/state as collapsed",
+                            default=False)
+
+    nosync = Property(dtype=bool,
+                      default=False,
+                      desc="Do not synchronize at the end of the state")
+
+    instrument = Property(
+        choices=dtypes.InstrumentationType,
+        desc="Measure execution statistics with given method",
+        default=dtypes.InstrumentationType.No_Instrumentation)
+
+    location = DictProperty(
+        key_type=str,
+        value_type=symbolic.pystr_to_symbolic,
+        desc='Full storage location identifier (e.g., rank, GPU ID)')
+
+    def __init__(self, label=None, sdfg=None, debuginfo=None, location=None):
+        """ Constructs an SDFG state.
+            :param label: Name for the state (optional).
+            :param sdfg: A reference to the parent SDFG.
+            :param debuginfo: Source code locator for debugging.
+        """
+        super(SDFGState, self).__init__()
+        self._label = label
+        self._parent = sdfg
+        self._graph = self  # Allowing MemletTrackingView mixin to work
+        self._clear_scopedict_cache()
+        self._debuginfo = debuginfo
+        self.is_collapsed = False
+        self.nosync = False
+        self.location = location if location is not None else {}
+
+    @property
+    def parent(self):
+        """ Returns the parent SDFG of this state. """
+        return self._parent
+
+    def __str__(self):
+        return self._label
+
+    # Clears the cached results for the scope_dict function.
+    # For use when the graph mutates (e.g., new edges/nodes, deletions)
+    def _clear_scopedict_cache(self):
+        self._scope_dict_toparent_cached = None
+        self._scope_dict_tochildren_cached = None
+        self._scope_tree_cached = None
+        self._scope_leaves_cached = None
+
+    @property
+    def label(self):
+        return self._label
+
+    @property
+    def name(self):
+        return self._label
+
+    def set_label(self, label):
+        self._label = label
+
+    def replace(self, name: str, new_name: str):
+        """ Finds and replaces all occurrences of a symbol or array in this
+            state.
+            :param name: Name to find.
+            :param new_name: Name to replace.
+        """
+        from dace.sdfg.sdfg import replace
+        replace(self, name, new_name)
+
+    def nodes(self) -> List[nd.Node]:
+        return super().nodes()
+
+    def add_node(self, node):
+        if not isinstance(node, nd.Node):
+            raise TypeError("Expected Node, got " + str(type(node)) + " (" +
+                            str(node) + ")")
+        self._clear_scopedict_cache()
+        return super(SDFGState, self).add_node(node)
+
+    def remove_node(self, node):
+        self._clear_scopedict_cache()
+        super(SDFGState, self).remove_node(node)
+
+    def add_edge(self, u, u_connector, v, v_connector, memlet):
+        if not isinstance(u, nd.Node):
+            raise TypeError("Source node is not of type nd.Node (type: %s)" %
+                            str(type(u)))
+        if u_connector is not None and not isinstance(u_connector, str):
+            raise TypeError("Source connector is not string (type: %s)" %
+                            str(type(u_connector)))
+        if not isinstance(v, nd.Node):
+            raise TypeError("Destination node is not of type nd.Node (type: " +
+                            "%s)" % str(type(v)))
+        if v_connector is not None and not isinstance(v_connector, str):
+            raise TypeError("Destination connector is not string (type: %s)" %
+                            str(type(v_connector)))
+        if not isinstance(memlet, mm.Memlet):
+            raise TypeError("Memlet is not of type Memlet (type: %s)" %
+                            str(type(memlet)))
+
+        self._clear_scopedict_cache()
+        return super(SDFGState, self).add_edge(u, u_connector, v, v_connector,
+                                               memlet)
+
+    def remove_edge(self, edge):
+        self._clear_scopedict_cache()
+        super(SDFGState, self).remove_edge(edge)
+
+    def remove_edge_and_connectors(self, edge):
+        self._clear_scopedict_cache()
+        super(SDFGState, self).remove_edge(edge)
+        if edge.src_conn in edge.src.out_connectors:
+            edge.src._out_connectors.remove(edge.src_conn)
+        if edge.dst_conn in edge.dst.in_connectors:
+            edge.dst._in_connectors.remove(edge.dst_conn)
+
+    def all_nodes_recursive(self):
+        for node in self.nodes():
+            yield node, self
+            if isinstance(node, nd.NestedSDFG):
+                yield from node.sdfg.all_nodes_recursive()
+
+    def all_edges_recursive(self):
+        for e in self.edges():
+            yield e, self
+        for node in self.nodes():
+            if isinstance(node, nd.NestedSDFG):
+                yield from node.sdfg.all_edges_recursive()
+
+    def data_symbols(self):
+        """ Returns all symbols used in data nodes. """
+        from dace.sdfg.sdfg import data_symbols
+        return data_symbols(self)
+
+    def scope_symbols(self):
+        """ Returns all symbols defined by scopes within this state. """
+        from dace.sdfg.sdfg import scope_symbols
+        return scope_symbols(self)
+
+    def undefined_symbols(self, sdfg, include_scalar_data):
+        from dace.sdfg.sdfg import undefined_symbols
+        return undefined_symbols(sdfg, self, include_scalar_data)
+
+    def data_nodes(self):
+        """ Returns all data_nodes (arrays) present in this state. """
+        return [n for n in self.nodes() if isinstance(n, nd.AccessNode)]
+
+    def memlets_for_array(self, arrayname):
+        return [e for e in self.edges() if e[3].data == arrayname]
+
+    def to_json(self, parent=None):
+        # Create scope dictionary with a failsafe
+        try:
+            scope_dict = {
+                k: sorted(v)
+                for k, v in sorted(
+                    self.scope_dict(node_to_children=True,
+                                    return_ids=True).items())
+            }
+        except RuntimeError:
+            scope_dict = {}
+
+        ret = {
+            'type':
+            type(self).__name__,
+            'label':
+            self.name,
+            'id':
+            parent.node_id(self) if parent is not None else None,
+            'collapsed':
+            self.is_collapsed,
+            'scope_dict':
+            scope_dict,
+            'nodes': [n.to_json(self) for n in self.nodes()],
+            'edges': [
+                e.to_json(self) for e in sorted(
+                    self.edges(),
+                    key=lambda e: (e.src_conn or '', e.dst_conn or ''))
+            ],
+            'attributes':
+            serialize.all_properties_to_json(self),
+        }
+
+        return ret
+
+    @classmethod
+    def from_json(cls, json_obj, context={'sdfg': None}):
+        """ Loads the node properties, label and type into a dict.
+            :param json_obj: The object containing information about this node.
+                             NOTE: This may not be a string!
+            :return: An SDFGState instance constructed from the passed data
+        """
+
+        _type = json_obj['type']
+        if _type != cls.__name__:
+            raise Exception("Class type mismatch")
+
+        attrs = json_obj['attributes']
+        nodes = json_obj['nodes']
+        edges = json_obj['edges']
+
+        ret = SDFGState(label=json_obj['label'],
+                        sdfg=context['sdfg'],
+                        debuginfo=None)
+
+        rec_ci = {
+            'sdfg': context['sdfg'],
+            'sdfg_state': ret,
+            'callback': context['callback'] if 'callback' in context else None
+        }
+        serialize.set_properties_from_json(ret, json_obj, rec_ci)
+
+        for n in nodes:
+            nret = serialize.loads(serialize.dumps(n), context=rec_ci)
+            ret.add_node(nret)
+
+        # Connect using the edges
+        for e in edges:
+            eret = serialize.loads(serialize.dumps(e), context=rec_ci)
+
+            ret.add_edge(eret.src, eret.src_conn, eret.dst, eret.dst_conn,
+                         eret.data)
+
+        # Fix potentially broken scopes
+        for n in nodes:
+            if isinstance(n, nd.MapExit):
+                n.map = ret.entry_node(n).map
+            elif isinstance(n, nd.ConsumeExit):
+                n.consume = ret.entry_node(n).consume
+
+        return ret
+
+    def _repr_html_(self):
+        """ HTML representation of a state, used mainly for Jupyter
+            notebooks. """
+        # Create dummy SDFG with this state as the only one
+        from dace.sdfg import SDFG
+        arrays = set(n.data for n in self.data_nodes())
+        sdfg = SDFG(self.label)
+        sdfg._arrays = {k: self._parent.arrays[k] for k in arrays}
+        sdfg.add_node(self)
+
+        return sdfg._repr_html_()
+
+    # TODO: Move to scope.py
+    def scope_tree(self):
+        from dace.sdfg.scope import ScopeTree
+
+        if (hasattr(self, '_scope_tree_cached')
+                and self._scope_tree_cached is not None):
+            return copy.copy(self._scope_tree_cached)
+
+        sdp = self.scope_dict(node_to_children=False)
+        sdc = self.scope_dict(node_to_children=True)
+
+        result = {}
+
+        sdfg_symbols = self.parent.undefined_symbols(True).keys()
+
+        # Get scopes
+        for node, scopenodes in sdc.items():
+            if node is None:
+                exit_node = None
+            else:
+                exit_node = next(v for v in scopenodes
+                                 if isinstance(v, nd.ExitNode))
+            scope = ScopeTree(node, exit_node)
+            scope.defined_vars = set(
+                symbolic.pystr_to_symbolic(s)
+                for s in (self.parent.symbols_defined_at(node, self).keys()
+                          | sdfg_symbols))
+            result[node] = scope
+
+        # Scope parents and children
+        for node, scope in result.items():
+            if node is not None:
+                scope.parent = result[sdp[node]]
+            scope.children = [
+                result[n] for n in sdc[node] if isinstance(n, nd.EntryNode)
+            ]
+
+        self._scope_tree_cached = result
+
+        return copy.copy(self._scope_tree_cached)
+
+    def scope_leaves(self):
+        if (hasattr(self, '_scope_leaves_cached')
+                and self._scope_leaves_cached is not None):
+            return copy.copy(self._scope_leaves_cached)
+        st = self.scope_tree()
+        self._scope_leaves_cached = [
+            scope for scope in st.values() if len(scope.children) == 0
+        ]
+        return copy.copy(self._scope_leaves_cached)
+
+    def scope_dict(self,
+                   node_to_children=False,
+                   return_ids=False,
+                   validate=True):
+        """ Returns a dictionary that segments an SDFG state into
+            entry-node/exit-node scopes.
+
+            :param node_to_children: If False (default), returns a mapping
+                                     of each node to its parent scope
+                                     (ScopeEntry) node. If True, returns a
+                                     mapping of each parent node to a list of
+                                     children nodes.
+            :type node_to_children: bool
+            :param return_ids: Return node ID numbers instead of node objects.
+            :type return_ids: bool
+            :param validate: Ensure that the graph is not malformed when
+                             computing dictionary.
+            :return: The mapping from a node to its parent scope node, or the
+                     mapping from a node to a list of children nodes.
+            :rtype: dict(Node, Node) or dict(Node, list(Node))
+        """
+        from dace.sdfg.scope import _scope_dict_inner, _scope_dict_to_ids
+        result = None
+        if not node_to_children and self._scope_dict_toparent_cached is not None:
+            result = copy.copy(self._scope_dict_toparent_cached)
+        elif node_to_children and self._scope_dict_tochildren_cached is not None:
+            result = copy.copy(self._scope_dict_tochildren_cached)
+
+        if result is None:
+            result = {}
+            node_queue = collections.deque(self.source_nodes())
+            eq = _scope_dict_inner(self, node_queue, None, node_to_children,
+                                   result)
+
+            # Sanity check
+            if validate and len(eq) != 0:
+                raise RuntimeError("Leftover nodes in queue: {}".format(eq))
+
+            # Cache result
+            if node_to_children:
+                self._scope_dict_tochildren_cached = result
+            else:
+                self._scope_dict_toparent_cached = result
+
+            result = copy.copy(result)
+
+        if return_ids:
+            return _scope_dict_to_ids(self, result)
+        return result
+
+    def scope_subgraph(self,
+                       entry_node,
+                       include_entry=True,
+                       include_exit=True):
+        from dace.sdfg.scope import _scope_subgraph
+        return _scope_subgraph(self, entry_node, include_entry, include_exit)
+
+    def top_level_transients(self):
+        """Iterate over top-level transients of this state."""
+        return top_level_transients(self)  # Free function
+
+    def all_transients(self):
+        """Iterate over all transients in this state."""
+        return all_transients(self)
+
+    def entry_node(self, node: nd.Node) -> nd.EntryNode:
+        """ Returns the entry node that wraps the current node, or None if
+            it is top-level in a state. """
+        return self.scope_dict()[node]
+
+    def exit_node(self, entry_node: nd.EntryNode) -> nd.ExitNode:
+        """ Returns the exit node leaving the context opened by
+            the given entry node. """
+        node_to_children = self.scope_dict(True)
+        return next(v for v in node_to_children[entry_node]
+                    if isinstance(v, nd.ExitNode))
+
+    # Dynamic SDFG creation API
+    ##############################
+    def add_read(self,
+                 array_or_stream_name: str,
+                 debuginfo=None) -> nd.AccessNode:
+        """ Adds a read-only access node to this SDFG state.
+            :param array_or_stream_name: The name of the array/stream.
+            :return: An array access node.
+        """
+        debuginfo = _getdebuginfo(debuginfo)
+        node = nd.AccessNode(array_or_stream_name,
+                             dtypes.AccessType.ReadOnly,
+                             debuginfo=debuginfo)
+        self.add_node(node)
+        return node
+
+    def add_write(self,
+                  array_or_stream_name: str,
+                  debuginfo=None) -> nd.AccessNode:
+        """ Adds a write-only access node to this SDFG state.
+            :param array_or_stream_name: The name of the array/stream.
+            :return: An array access node.
+        """
+        debuginfo = _getdebuginfo(debuginfo)
+        node = nd.AccessNode(array_or_stream_name,
+                             dtypes.AccessType.WriteOnly,
+                             debuginfo=debuginfo)
+        self.add_node(node)
+        return node
+
+    def add_access(self,
+                   array_or_stream_name: str,
+                   debuginfo=None) -> nd.AccessNode:
+        """ Adds a general (read/write) access node to this SDFG state.
+            :param array_or_stream_name: The name of the array/stream.
+            :return: An array access node.
+        """
+        debuginfo = _getdebuginfo(debuginfo)
+        node = nd.AccessNode(array_or_stream_name,
+                             dtypes.AccessType.ReadWrite,
+                             debuginfo=debuginfo)
+        self.add_node(node)
+        return node
+
+    def add_tasklet(
+        self,
+        name: str,
+        inputs: Set[str],
+        outputs: Set[str],
+        code: str,
+        language: dtypes.Language = dtypes.Language.Python,
+        location: dict = None,
+        debuginfo=None,
+    ):
+        """ Adds a tasklet to the SDFG state. """
+        debuginfo = _getdebuginfo(debuginfo)
+        tasklet = nd.Tasklet(
+            name,
+            inputs,
+            outputs,
+            code,
+            language,
+            location=location,
+            debuginfo=debuginfo,
+        )
+        self.add_node(tasklet)
+        return tasklet
+
+    def add_nested_sdfg(
+        self,
+        sdfg: 'dace.sdfg.SDFG',
+        parent,
+        inputs: Set[str],
+        outputs: Set[str],
+        symbol_mapping: Dict[str, Any] = None,
+        name=None,
+        schedule=dtypes.ScheduleType.Default,
+        location=None,
+        debuginfo=None,
+    ):
+        """ Adds a nested SDFG to the SDFG state. """
+        if name is None:
+            name = sdfg.label
+        debuginfo = _getdebuginfo(debuginfo)
+
+        sdfg.parent = self
+        sdfg._parent_sdfg = self.parent
+
+        sdfg.update_sdfg_list([])
+
+        s = nd.NestedSDFG(
+            name,
+            sdfg,
+            inputs,
+            outputs,
+            symbol_mapping=symbol_mapping,
+            schedule=schedule,
+            location=location,
+            debuginfo=debuginfo,
+        )
+        self.add_node(s)
+
+        # Add "default" undefined symbols if None are given
+        symbols = sdfg.undefined_symbols(False)
+        if symbol_mapping is None:
+            symbol_mapping = {s: s for s in symbols.keys()}
+            s.symbol_mapping = symbol_mapping
+
+        # Validate missing symbols
+        missing_symbols = [
+            s for s in symbols.keys() if s not in symbol_mapping
+        ]
+        if missing_symbols:
+            raise ValueError('Missing symbols on nested SDFG "%s": %s' %
+                             (name, missing_symbols))
+
+        return s
+
+    def _map_from_ndrange(self,
+                          name,
+                          schedule,
+                          unroll,
+                          ndrange,
+                          debuginfo=None):
+        # Input can either be a dictionary or a list of pairs
+        if isinstance(ndrange, list):
+            params = [k for k, v in ndrange]
+            ndrange = {k: v for k, v in ndrange}
+        else:
+            params = list(ndrange.keys())
+
+        map_range = SubsetProperty.from_string(", ".join(
+            [ndrange[p] for p in params]))
+        map = nd.Map(name,
+                     params,
+                     map_range,
+                     schedule,
+                     unroll,
+                     debuginfo=debuginfo)
+        return map
+
+    def add_map(
+            self,
+            name,
+            ndrange: Union[Dict[str, str], List[Tuple[str, str]]],
+            schedule=dtypes.ScheduleType.Default,
+            unroll=False,
+            debuginfo=None,
+    ) -> Tuple[nd.MapEntry, nd.MapExit]:
+        """ Adds a map entry and map exit.
+            :param name:      Map label
+            :param ndrange:   Mapping between range variable names and their
+                              subsets (parsed from strings)
+            :param schedule:  Map schedule type
+            :param unroll:    True if should unroll the map in code generation
+
+            :return: (map_entry, map_exit) node 2-tuple
+        """
+        debuginfo = _getdebuginfo(debuginfo)
+        map = self._map_from_ndrange(name,
+                                     schedule,
+                                     unroll,
+                                     ndrange,
+                                     debuginfo=debuginfo)
+        map_entry = nd.MapEntry(map)
+        map_exit = nd.MapExit(map)
+        self.add_nodes_from([map_entry, map_exit])
+        return map_entry, map_exit
+
+    def add_consume(
+        self,
+        name,
+        elements: Tuple[str, str],
+        condition: str = None,
+        schedule=dtypes.ScheduleType.Default,
+        chunksize=1,
+        debuginfo=None,
+        language=dtypes.Language.Python
+    ) -> Tuple[nd.ConsumeEntry, nd.ConsumeExit]:
+        """ Adds consume entry and consume exit nodes.
+            :param name:      Label
+            :param elements:  A 2-tuple signifying the processing element
+                              index and number of total processing elements
+            :param condition: Quiescence condition to finish consuming, or
+                              None (by default) to consume until the stream
+                              is empty for the first time. If false, will
+                              consume forever.
+            :param schedule:  Consume schedule type.
+            :param chunksize: Maximal number of elements to consume at a time.
+            :param debuginfo: Source code line information for debugging.
+            :param language:  Code language for ``condition``.
+
+            :return: (consume_entry, consume_exit) node 2-tuple
+        """
+        if len(elements) != 2:
+            raise TypeError("Elements must be a 2-tuple of "
+                            "(PE_index, num_PEs)")
+        pe_tuple = (elements[0], SymbolicProperty.from_string(elements[1]))
+
+        debuginfo = _getdebuginfo(debuginfo)
+        consume = nd.Consume(name,
+                             pe_tuple,
+                             CodeBlock(condition, language),
+                             schedule,
+                             chunksize,
+                             debuginfo=debuginfo)
+        entry = nd.ConsumeEntry(consume)
+        exit = nd.ConsumeExit(consume)
+
+        self.add_nodes_from([entry, exit])
+        return entry, exit
+
+    def add_mapped_tasklet(
+            self,
+            name: str,
+            map_ranges: Dict[str, sbs.Subset],
+            inputs: Dict[str, mm.Memlet],
+            code: str,
+            outputs: Dict[str, mm.Memlet],
+            schedule=dtypes.ScheduleType.Default,
+            unroll_map=False,
+            location=None,
+            language=dtypes.Language.Python,
+            debuginfo=None,
+            external_edges=False,
+            input_nodes: Optional[Dict[str, nd.AccessNode]] = None,
+            output_nodes: Optional[Dict[str, nd.AccessNode]] = None,
+            propagate=True) -> Tuple[nd.Tasklet, nd.MapEntry, nd.MapExit]:
+        """ Convenience function that adds a map entry, tasklet, map exit,
+            and the respective edges to external arrays.
+            :param name:       Tasklet (and wrapping map) name
+            :param map_ranges: Mapping between variable names and their
+                               subsets
+            :param inputs:     Mapping between input local variable names and
+                               their memlets
+            :param code:       Code (written in `language`)
+            :param outputs:    Mapping between output local variable names and
+                               their memlets
+            :param schedule:   Map schedule
+            :param unroll_map: True if map should be unrolled in code
+                               generation
+            :param location:   Execution location indicator.
+            :param language:   Programming language in which the code is
+                               written
+            :param debuginfo:  Debugging information (mostly for DIODE)
+            :param external_edges: Create external access nodes and connect
+                                   them with memlets automatically
+            :param input_nodes: Mapping between data names and corresponding
+                                input nodes to link to, if external_edges is
+                                True.
+            :param output_nodes: Mapping between data names and corresponding
+                                 output nodes to link to, if external_edges is
+                                 True.
+            :param propagate: If True, computes outer memlets via propagation.
+                              False will run faster but the SDFG may not be
+                              semantically correct.
+            :return: tuple of (tasklet, map_entry, map_exit)
+        """
+        map_name = name + "_map"
+        debuginfo = _getdebuginfo(debuginfo)
+        tasklet = nd.Tasklet(
+            name,
+            set(inputs.keys()),
+            set(outputs.keys()),
+            code,
+            language=language,
+            location=location,
+            debuginfo=debuginfo,
+        )
+        map = self._map_from_ndrange(map_name,
+                                     schedule,
+                                     unroll_map,
+                                     map_ranges,
+                                     debuginfo=debuginfo)
+        map_entry = nd.MapEntry(map)
+        map_exit = nd.MapExit(map)
+        self.add_nodes_from([map_entry, tasklet, map_exit])
+
+        # Create access nodes
+        inpdict = {}
+        outdict = {}
+        if external_edges:
+            input_nodes = input_nodes or {}
+            output_nodes = output_nodes or {}
+            input_data = set(memlet.data for memlet in inputs.values())
+            output_data = set(memlet.data for memlet in outputs.values())
+            for inp in input_data:
+                if inp in input_nodes:
+                    inpdict[inp] = input_nodes[inp]
+                else:
+                    inpdict[inp] = self.add_read(inp)
+            for out in output_data:
+                if out in output_nodes:
+                    outdict[out] = output_nodes[out]
+                else:
+                    outdict[out] = self.add_write(out)
+
+        # Connect inputs from map to tasklet
+        tomemlet = {}
+        for name, memlet in inputs.items():
+            # Set memlet local name
+            memlet.name = name
+            # Add internal memlet edge
+            self.add_edge(map_entry, None, tasklet, name, memlet)
+            tomemlet[memlet.data] = memlet
+
+        # If there are no inputs, add empty memlet
+        if len(inputs) == 0:
+            self.add_edge(map_entry, None, tasklet, None, mm.EmptyMemlet())
+
+        if external_edges:
+            for inp, inpnode in inpdict.items():
+                # Add external edge
+                if propagate:
+                    outer_memlet = propagate_memlet(self, tomemlet[inp],
+                                                    map_entry, True)
+                else:
+                    outer_memlet = tomemlet[inp]
+                self.add_edge(inpnode, None, map_entry, "IN_" + inp,
+                              outer_memlet)
+
+                # Add connectors to internal edges
+                for e in self.out_edges(map_entry):
+                    if e.data.data == inp:
+                        e._src_conn = "OUT_" + inp
+
+                # Add connectors to map entry
+                map_entry.add_in_connector("IN_" + inp)
+                map_entry.add_out_connector("OUT_" + inp)
+
+        # Connect outputs from tasklet to map
+        tomemlet = {}
+        for name, memlet in outputs.items():
+            # Set memlet local name
+            memlet.name = name
+            # Add internal memlet edge
+            self.add_edge(tasklet, name, map_exit, None, memlet)
+            tomemlet[memlet.data] = memlet
+
+        # If there are no outputs, add empty memlet
+        if len(outputs) == 0:
+            self.add_edge(tasklet, None, map_exit, None, mm.EmptyMemlet())
+
+        if external_edges:
+            for out, outnode in outdict.items():
+                # Add external edge
+                if propagate:
+                    outer_memlet = propagate_memlet(self, tomemlet[out],
+                                                    map_exit, True)
+                else:
+                    outer_memlet = tomemlet[out]
+                self.add_edge(map_exit, "OUT_" + out, outnode, None,
+                              outer_memlet)
+
+                # Add connectors to internal edges
+                for e in self.in_edges(map_exit):
+                    if e.data.data == out:
+                        e._dst_conn = "IN_" + out
+
+                # Add connectors to map entry
+                map_exit.add_in_connector("IN_" + out)
+                map_exit.add_out_connector("OUT_" + out)
+
+        return tasklet, map_entry, map_exit
+
+    def add_reduce(
+            self,
+            wcr,
+            axes,
+            identity=None,
+            schedule=dtypes.ScheduleType.Default,
+            debuginfo=None,
+    ) -> 'dace.libraries.standard.Reduce':
+        """ Adds a reduction node.
+            :param wcr: A lambda function representing the reduction operation
+            :param axes: A tuple of axes to reduce the input memlet from, or
+                         None for all axes
+            :param identity: If not None, initializes output memlet values
+                                 with this value
+            :param schedule: Reduction schedule type
+
+            :return: A Reduce node
+        """
+        import dace.libraries.standard as stdlib  # Avoid import loop
+        debuginfo = _getdebuginfo(debuginfo)
+        result = stdlib.Reduce(wcr,
+                               axes,
+                               identity,
+                               schedule=schedule,
+                               debuginfo=debuginfo)
+        self.add_node(result)
+        return result
+
+    def add_edge_pair(
+        self,
+        scope_node,
+        internal_node,
+        external_node,
+        internal_memlet,
+        external_memlet=None,
+        scope_connector=None,
+        internal_connector=None,
+        external_connector=None,
+    ):
+        """ Adds two edges around a scope node (e.g., map entry, consume
+            exit).
+
+            The internal memlet (connecting to the internal node) has to be
+            specified. If external_memlet (i.e., connecting to the node out
+            of the scope) is not specified, it is propagated automatically
+            using internal_memlet and the scope.
+
+            :param scope_node: A scope node (for example, map exit) to add
+                               edges around.
+            :param internal_node: The node within the scope to connect to. If
+                                  `scope_node` is an entry node, this means
+                                  the node connected to the outgoing edge,
+                                  else incoming.
+            :param external_node: The node out of the scope to connect to.
+            :param internal_memlet: The memlet on the edge to/from
+                                    internal_node.
+            :param external_memlet: The memlet on the edge to/from
+                                    external_node (optional, will propagate
+                                    internal_memlet if not specified).
+            :param scope_connector: A scope connector name (or a unique
+                                    number if not specified).
+            :param internal_connector: The connector on internal_node to
+                                       connect to.
+            :param external_connector: The connector on external_node to
+                                       connect to.
+            :return: A 2-tuple representing the (internal, external) edges.
+        """
+        if not isinstance(scope_node, (nd.EntryNode, nd.ExitNode)):
+            raise ValueError("scope_node is not a scope entry/exit")
+
+        # Autodetermine scope connector ID
+        if scope_connector is None:
+            # Pick out numbered connectors that do not lead into the scope range
+            conn_id = 1
+            for conn in scope_node.in_connectors | scope_node.out_connectors:
+                if conn.startswith("IN_") or conn.startswith("OUT_"):
+                    conn_name = conn[conn.find("_") + 1:]
+                    try:
+                        cid = int(conn_name)
+                        if cid >= conn_id:
+                            conn_id = cid + 1
+                    except (TypeError, ValueError):
+                        pass
+            scope_connector = str(conn_id)
+
+        # Add connectors
+        scope_node.add_in_connector("IN_" + scope_connector)
+        scope_node.add_out_connector("OUT_" + scope_connector)
+        ##########################
+
+        # Add internal edge
+        if isinstance(scope_node, nd.EntryNode):
+            iedge = self.add_edge(
+                scope_node,
+                "OUT_" + scope_connector,
+                internal_node,
+                internal_connector,
+                internal_memlet,
+            )
+        else:
+            iedge = self.add_edge(
+                internal_node,
+                internal_connector,
+                scope_node,
+                "IN_" + scope_connector,
+                internal_memlet,
+            )
+
+        # Add external edge
+        if external_memlet is None:
+            # If undefined, propagate
+            external_memlet = propagate_memlet(self, internal_memlet,
+                                               scope_node, True)
+
+        if isinstance(scope_node, nd.EntryNode):
+            eedge = self.add_edge(
+                external_node,
+                external_connector,
+                scope_node,
+                "IN_" + scope_connector,
+                external_memlet,
+            )
+        else:
+            eedge = self.add_edge(
+                scope_node,
+                "OUT_" + scope_connector,
+                external_node,
+                external_connector,
+                external_memlet,
+            )
+
+        return (iedge, eedge)
+
+    def add_memlet_path(self,
+                        *path_nodes,
+                        memlet=None,
+                        src_conn=None,
+                        dst_conn=None,
+                        propagate=True):
+        """ Adds a path of memlet edges between the given nodes, propagating
+            from the given innermost memlet.
+
+            :param *path_nodes: Nodes participating in the path (in the given
+                                order).
+            :keyword memlet: (mandatory) The memlet at the innermost scope
+                             (e.g., the incoming memlet to a tasklet (last
+                             node), or an outgoing memlet from an array
+                             (first node), followed by scope exits).
+            :keyword src_conn: Connector at the beginning of the path.
+            :keyword dst_conn: Connector at the end of the path.
+        """
+        if memlet is None:
+            raise TypeError("Innermost memlet cannot be None")
+        if len(path_nodes) < 2:
+            raise ValueError("Memlet path must consist of at least 2 nodes")
+
+        src_node = path_nodes[0]
+        dst_node = path_nodes[-1]
+
+        # Add edges first so that scopes can be understood
+        edges = [
+            self.add_edge(path_nodes[i], None, path_nodes[i + 1], None,
+                          mm.EmptyMemlet())
+            for i in range(len(path_nodes) - 1)
+        ]
+
+        if not isinstance(memlet, mm.Memlet):
+            raise TypeError("Expected Memlet, got: {}".format(
+                type(memlet).__name__))
+
+        if any(isinstance(n, nd.EntryNode) for n in path_nodes):
+            propagate_forward = False
+        else:  # dst node's scope is higher than src node, propagate out
+            propagate_forward = True
+
+        # Innermost edge memlet
+        cur_memlet = memlet
+
+        # Verify that connectors exist
+        if (not isinstance(memlet, mm.EmptyMemlet)
+                and hasattr(edges[0].src, "out_connectors")
+                and isinstance(edges[0].src, nd.CodeNode)
+                and not isinstance(edges[0].src, nd.LibraryNode) and
+            (src_conn is None or src_conn not in edges[0].src.out_connectors)):
+            raise ValueError("Output connector {} does not exist in {}".format(
+                src_conn, edges[0].src.label))
+        if (not isinstance(memlet, mm.EmptyMemlet)
+                and hasattr(edges[-1].dst, "in_connectors")
+                and isinstance(edges[-1].dst, nd.CodeNode)
+                and not isinstance(edges[-1].dst, nd.LibraryNode) and
+            (dst_conn is None or dst_conn not in edges[-1].dst.in_connectors)):
+            raise ValueError("Input connector {} does not exist in {}".format(
+                dst_conn, edges[-1].dst.label))
+
+        path = edges if propagate_forward else reversed(edges)
+        # Propagate and add edges
+        for i, edge in enumerate(path):
+            # Figure out source and destination connectors
+            if propagate_forward:
+                sconn = src_conn if i == 0 else ("OUT_" +
+                                                 edge.src.last_connector())
+                dconn = (dst_conn if i == len(edges) - 1 else
+                         ("IN_" + edge.dst.next_connector()))
+            else:
+                sconn = (src_conn if i == len(edges) - 1 else
+                         ("OUT_" + edge.src.next_connector()))
+                dconn = dst_conn if i == 0 else ("IN_" +
+                                                 edge.dst.last_connector())
+
+            if isinstance(cur_memlet, mm.EmptyMemlet):
+                if propagate_forward:
+                    sconn = src_conn if i == 0 else None
+                    dconn = dst_conn if i == len(edges) - 1 else None
+                else:
+                    sconn = src_conn if i == len(edges) - 1 else None
+                    dconn = dst_conn if i == 0 else None
+
+            # Modify edge to match memlet path
+            edge._src_conn = sconn
+            edge._dst_conn = dconn
+            edge._data = cur_memlet
+
+            # Add connectors to edges
+            if propagate_forward:
+                if dconn is not None:
+                    edge.dst.add_in_connector(dconn)
+                if sconn is not None:
+                    edge.src.add_out_connector(sconn)
+            else:
+                if dconn is not None:
+                    edge.dst.add_in_connector(dconn)
+                if sconn is not None:
+                    edge.src.add_out_connector(sconn)
+
+            # Propagate current memlet to produce the next one
+            if i < len(edges) - 1:
+                snode = edge.dst if propagate_forward else edge.src
+                if not isinstance(cur_memlet, mm.EmptyMemlet):
+                    if propagate:
+                        cur_memlet = propagate_memlet(self, cur_memlet, snode,
+                                                      True)
+
+    # DEPRECATED FUNCTIONS
+    ######################################
+    def add_array(self,
+                  name,
+                  shape,
+                  dtype,
+                  storage=dtypes.StorageType.Default,
+                  materialize_func=None,
+                  transient=False,
+                  strides=None,
+                  offset=None,
+                  lifetime=dtypes.AllocationLifetime.Scope,
+                  debuginfo=None,
+                  total_size=None,
+                  find_new_name=False,
+                  alignment=0):
+        """ @attention: This function is deprecated. """
+        warnings.warn(
+            'The "SDFGState.add_array" API is deprecated, please '
+            'use "SDFG.add_array" and "SDFGState.add_access"',
+            DeprecationWarning)
+        # Workaround to allow this legacy API
+        if name in self.parent._arrays:
+            del self.parent._arrays[name]
+        self.parent.add_array(name,
+                              shape,
+                              dtype,
+                              storage,
+                              materialize_func,
+                              transient,
+                              strides,
+                              offset,
+                              lifetime,
+                              debuginfo,
+                              find_new_name=find_new_name,
+                              total_size=total_size,
+                              alignment=alignment)
+        return self.add_access(name, debuginfo)
+
+    def add_stream(
+        self,
+        name,
+        dtype,
+        veclen=1,
+        buffer_size=1,
+        shape=(1, ),
+        storage=dtypes.StorageType.Default,
+        transient=False,
+        offset=None,
+        lifetime=dtypes.AllocationLifetime.Scope,
+        debuginfo=None,
+    ):
+        """ @attention: This function is deprecated. """
+        warnings.warn(
+            'The "SDFGState.add_stream" API is deprecated, please '
+            'use "SDFG.add_stream" and "SDFGState.add_access"',
+            DeprecationWarning)
+        # Workaround to allow this legacy API
+        if name in self.parent._arrays:
+            del self.parent._arrays[name]
+        self.parent.add_stream(
+            name,
+            dtype,
+            veclen,
+            buffer_size,
+            shape,
+            storage,
+            transient,
+            offset,
+            lifetime,
+            debuginfo,
+        )
+        return self.add_access(name, debuginfo)
+
+    def add_scalar(
+        self,
+        name,
+        dtype,
+        storage=dtypes.StorageType.Default,
+        transient=False,
+        lifetime=dtypes.AllocationLifetime.Scope,
+        debuginfo=None,
+    ):
+        """ @attention: This function is deprecated. """
+        warnings.warn(
+            'The "SDFGState.add_scalar" API is deprecated, please '
+            'use "SDFG.add_scalar" and "SDFGState.add_access"',
+            DeprecationWarning)
+        # Workaround to allow this legacy API
+        if name in self.parent._arrays:
+            del self.parent._arrays[name]
+        self.parent.add_scalar(name, dtype, storage, transient, lifetime,
+                               debuginfo)
+        return self.add_access(name, debuginfo)
+
+    def add_transient(self,
+                      name,
+                      shape,
+                      dtype,
+                      storage=dtypes.StorageType.Default,
+                      materialize_func=None,
+                      strides=None,
+                      offset=None,
+                      lifetime=dtypes.AllocationLifetime.Scope,
+                      debuginfo=None,
+                      total_size=None,
+                      alignment=0):
+        """ @attention: This function is deprecated. """
+        return self.add_array(name,
+                              shape,
+                              dtype,
+                              storage,
+                              materialize_func,
+                              True,
+                              strides,
+                              offset,
+                              lifetime,
+                              debuginfo,
+                              total_size=total_size,
+                              alignment=alignment)
+
+    # SDFG queries
+    ######################################
+    def find_node(self, node_id_or_label):
+        """ Finds a node according to its ID (if integer is
+            provided) or label (if string is provided).
+
+            :param node_id_or_label  Node ID (if int) or label (if str)
+            :return A nodes.Node object
+        """
+
+        if isinstance(node_id_or_label, str):
+            for n in self.nodes():
+                if n.label == node_id_or_label:
+                    return n
+            raise LookupError("Node %s not found" % node_id_or_label)
+        elif isinstance(node_id_or_label, int):
+            return self.nodes()[node_id_or_label]
+        else:
+            raise TypeError("node_id_or_label is not an int nor string")
+
+    def is_empty(self):
+        return len([n for n in self.nodes()]) == 0
+
+    def fill_scope_connectors(self):
+        """ Creates new "IN_%d" and "OUT_%d" connectors on each scope entry
+            and exit, depending on array names. """
+        for nid, node in enumerate(self.nodes()):
+            ####################################################
+            # Add connectors to scope entries
+            if isinstance(node, nd.EntryNode):
+                # Find current number of input connectors
+                num_inputs = len([
+                    e for e in self.in_edges(node)
+                    if e.dst_conn is not None and e.dst_conn.startswith("IN_")
+                ])
+
+                conn_to_data = {}
+
+                # Append input connectors and get mapping of connectors to data
+                for edge in self.in_edges(node):
+                    if edge.dst_conn is not None and edge.dst_conn.startswith(
+                            "IN_"):
+                        conn_to_data[edge.data.data] = edge.dst_conn[3:]
+
+                    # We're only interested in edges without connectors
+                    if edge.dst_conn is not None or edge.data.data is None:
+                        continue
+                    edge._dst_conn = "IN_" + str(num_inputs + 1)
+                    node._in_connectors.add(edge.dst_conn)
+                    conn_to_data[edge.data.data] = num_inputs + 1
+
+                    num_inputs += 1
+
+                # Set the corresponding output connectors
+                for edge in self.out_edges(node):
+                    if edge.src_conn is not None:
+                        continue
+                    if edge.data.data is None:
+                        continue
+                    edge._src_conn = "OUT_" + str(conn_to_data[edge.data.data])
+                    node._out_connectors.add(edge.src_conn)
+            ####################################################
+            # Same treatment for scope exits
+            if isinstance(node, nd.ExitNode):
+                # Find current number of output connectors
+                num_outputs = len([
+                    e for e in self.out_edges(node)
+                    if e.src_conn is not None and e.src_conn.startswith("OUT_")
+                ])
+
+                conn_to_data = {}
+
+                # Append output connectors and get mapping of connectors to data
+                for edge in self.out_edges(node):
+                    if edge.src_conn is not None and edge.src_conn.startswith(
+                            "OUT_"):
+                        conn_to_data[edge.data.data] = edge.src_conn[4:]
+
+                    # We're only interested in edges without connectors
+                    if edge.src_conn is not None or edge.data.data is None:
+                        continue
+                    edge._src_conn = "OUT_" + str(num_outputs + 1)
+                    node._out_connectors.add(edge.src_conn)
+                    conn_to_data[edge.data.data] = num_outputs + 1
+
+                    num_outputs += 1
+
+                # Set the corresponding input connectors
+                for edge in self.in_edges(node):
+                    if edge.dst_conn is not None:
+                        continue
+                    if edge.data.data is None:
+                        continue
+                    edge._dst_conn = "IN_" + str(conn_to_data[edge.data.data])
+                    node._in_connectors.add(edge.dst_conn)
+
+    def validate(self) -> None:
+        validate_state(self)
+
+
+def top_level_transients(dfg):
+    """ Iterate over top-level transients (i.e., ones that exist in multiple
+        states or scopes) of the passed dataflow graph. """
+    sdfg = dfg.parent
+    visited_transients = set()
+    scope_dict = dfg.scope_dict(node_to_children=True)
+    for node in scope_dict[None]:  # Top-level nodes
+        if not isinstance(node, nd.AccessNode):
+            continue
+        if node.data in visited_transients:
+            continue
+        if not node.desc(sdfg).transient:
+            continue
+        visited_transients.add(node.data)
+        yield node.data
+
+
+def all_transients(dfg):
+    """ Iterate over all transient data in the specified dataflow graph. """
+    visited = set()
+    for node in dfg.nodes():
+        if not isinstance(node, nd.AccessNode):
+            continue
+        if not node.desc(dfg.parent).transient:
+            continue
+        if node.data in visited:
+            continue
+        visited.add(node.data)
+        yield node.data
